Refactor (#7)
* Saving files before refreshing line endings

* Added .gitattributes file

* Bash file cleanup to work on windows.

Renamed the bash scripts under /docker/run to *.sh file format so they
can be picked up by .gitattribute. Also added a check in manage.sh for
windows environments; winpty command needs to be prepended to docker
command on windows in able for the container to work.

* Removed old apps to refactor into a cleaner app.

- The original apps were pretty disorganized, so we'll be starting a new
shift_tracking app with a cleaner folder structure to start fresh.

* Refactored to new folder structure.

The folder structure has been restructured to be more organized and maintainable. See App Structure page on Notion for style guide.